SMS-3993: Regenerated with PAL changes. Updated to Crypto v4 E2.

+++ apps/prime_apps/prime_dual_service_modem/firmware/src/config/pic32cx_mtg_ek_pl460_rf215/crypto/common_crypto/MCHP_Crypto_Aead_Cipher.h
@@ -71,7 +71,7 @@ typedef struct
     uint8_t *ptr_key;
     uint32_t aeadKeySize;
     uint8_t *ptr_initVect;
-    uint16_t initVectLen;    
+    uint32_t initVectLen;    
     uint8_t arr_aeadDataCtx[512]__attribute__((aligned (4)));
 }st_Crypto_Aead_AesGcm_ctx;
